requirements.txt
requests
tqdm
transformers
torch
torchvision
torchaudio
nltk
scikit-learn
Pillow
# PyTorch with CUDA support for GPU acceleration
--extra-index-url https://download.pytorch.org/whl/cu118
